requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.5.0
async-timeout==4.0.3
attrs==24.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.8.2
build==1.2.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
dataclasses-json==0.6.7
deepdiff==8.0.1
Deprecated==1.2.14
effdet==0.4.1
emoji==2.13.0
et-xmlfile==1.1.0
exceptiongroup==1.2.2
faiss-cpu==1.8.0.post1
filelock==3.16.1
filetype==1.2.0
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.9.0
gcloud>=0.17.0
google-api-core==2.20.0
google-auth==2.35.0
google-cloud-vision==3.7.4
googleapis-common-protos==1.65.0
grpcio==1.66.1
grpcio-status==1.66.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.25.0
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
iopath==0.1.10
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jws==0.1.3
kiwisolver==1.4.7
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.1
langchain-text-splitters==0.3.0
langchain-unstructured==0.1.4
langdetect==1.0.9
langsmith==0.1.123
layoutparser==0.3.4
lxml==5.3.0
Markdown==3.7
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.2
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.1
numpy==1.26.4
oauth2client>=3.0.0
olefile==0.47
ollama==0.3.3
omegaconf==2.3.0
onnx==1.16.2
onnxruntime==1.19.2
opencv-python==4.10.0.84
openpyxl==3.1.5
orderly-set==5.2.2
orjson==3.10.7
packaging==24.1
pandas==2.2.2
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.4
pi_heif==0.16.0
pikepdf==9.2.1
pillow==10.4.0
pip-tools==7.4.1
portalocker==2.10.1
proto-plus==1.24.0
protobuf==5.28.2
psutil==6.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycocotools==2.0.8
pycparser==2.22
pycryptodome>=3.6.4
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pyotp==2.9.0
pypandoc==1.13
pyparsing==3.1.4
pypdf==5.0.0
pypdfium2==4.30.0
pyproject_hooks==1.1.0
Pyrebase5==5.0.1

pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2024.4.27
python-jwt>=3.2.0
python-magic==0.4.27
python-multipart==0.0.9
python-oxmsg==0.0.1
python-pptx==1.0.2
pytz==2024.2
PyYAML==6.0.2
rapidfuzz==3.9.7
redis==5.0.8
regex==2024.9.11
requests>=2.31
requests-toolbelt==1.0.0
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.13.1
sentence-transformers==3.1.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
sympy==1.13.3
tabulate==0.9.0
tenacity==8.5.0
threadpoolctl==3.5.0
timm==1.0.9
tokenizers==0.19.1
tomli==2.0.1
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
transformers==4.44.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
unstructured==0.15.12
unstructured-client==0.25.9
unstructured-inference==0.7.36
unstructured.pytesseract==0.3.13
urllib3==1.26.20
Werkzeug==3.0.4
wrapt==1.16.0
xlrd==2.0.1
XlsxWriter==3.2.0
yarl==1.11.1
zipp==3.20.2
